requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
pyyaml
shap